+++ android/core/src/main/java/com/stadiamaps/ferrostar/core/NavigationViewModel.kt
@@ -6,15 +6,11 @@ import androidx.lifecycle.viewModelScope
 import com.stadiamaps.ferrostar.core.extensions.deviation
 import com.stadiamaps.ferrostar.core.extensions.progress
 import com.stadiamaps.ferrostar.core.extensions.visualInstruction
-import java.util.concurrent.Executors
-import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.flow.update
 import uniffi.ferrostar.GeographicCoordinate
-import uniffi.ferrostar.Heading
 import uniffi.ferrostar.RouteDeviation
 import uniffi.ferrostar.SpokenInstruction
 import uniffi.ferrostar.TripProgress
@@ -85,57 +81,6 @@ interface NavigationViewModel {
 
   // TODO: We think the camera may eventually need to be owned by the view model, but that's going
   // to be a very big refactor (maybe even crossing into the MapLibre Compose project)
-}
-
-class IdleNavigationViewModel(locationProvider: LocationProvider) :
-    ViewModel(), NavigationViewModel {
-  private val locationStateFlow = MutableStateFlow(locationProvider.lastLocation)
-  private val executor = Executors.newSingleThreadScheduledExecutor()
-
-  init {
-    locationProvider.addListener(
-        object : LocationUpdateListener {
-          override fun onLocationUpdated(location: UserLocation) {
-            locationStateFlow.update { location }
-          }
-
-          override fun onHeadingUpdated(heading: Heading) {
-            // TODO: Heading
-          }
-        },
-        executor)
-  }
-
-  override val uiState =
-      locationStateFlow
-          .map { userLocation ->
-            // TODO: Heading
-            NavigationUiState(userLocation, null, null, null, null, null, null, false, null, null)
-          }
-          .stateIn(
-              scope = viewModelScope,
-              started = SharingStarted.WhileSubscribed(),
-              // TODO: Heading
-              initialValue =
-                  NavigationUiState(
-                      locationProvider.lastLocation,
-                      null,
-                      null,
-                      null,
-                      null,
-                      null,
-                      null,
-                      false,
-                      null,
-                      null))
-
-  override fun toggleMute() {
-    // Do nothing
-  }
-
-  override fun stopNavigation() {
-    // Do nothing
-  }
 }
 
 class DefaultNavigationViewModel(
